Make Powergate CGO free & use testnet/3 (#210)
* progress pr


* fix docker embed


* adjust to lotus-client narrwoing scope


* update lotus-devnet docker image tag


* tons of progress towards testnet/3


* more changes, finally compiles


* fix tests due to testnet/3


* more changes for testnet/3 test passing


* change tests since testnet/3 devnet doesnt support more than one deal


* use timestamp for storage-ask


* fix ask test


* improve watching deals for deals


* disallow automatic restart on embed since stop might be indication of a problem, better this way


* prepare for pr


* update deps


* remove API interface from ask index. go mod tidy


* fix build commands in Readme


* remove GetCidInfo in favor to Show

+++ ffs/filcold/lotuschain/lotuschain.go
@@ -33,5 +33,5 @@ func (lc *LotusChain) GetHeight(ctx context.Context) (uint64, error) {
 	if err != nil {
 		return 0, fmt.Errorf("get head from lotus node: %s", err)
 	}
-	return h.Height(), nil
+	return uint64(h.Height()), nil
 }
